+++ c/tests/test_core.c
@@ -97,7 +97,7 @@ set_u64_le(uint8_t *dest, uint64_t value)
 }
 
 static void
-test_json_binary_metadata_get_blob(void)
+test_json_struct_metadata_get_blob(void)
 {
     int ret;
     char metadata[128];
@@ -132,7 +132,7 @@ test_json_binary_metadata_get_blob(void)
     memcpy(bytes + header_length, json_payload, json_length);
     memcpy(bytes + header_length + json_length, binary_payload, payload_length);
     metadata_length = (tsk_size_t) total_length;
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, 0);
     CU_ASSERT_PTR_EQUAL(json, (const char *) bytes + header_length);
@@ -149,7 +149,7 @@ test_json_binary_metadata_get_blob(void)
     CU_ASSERT_FATAL(total_length <= sizeof(metadata));
     set_u64_le(bytes + 13, (uint64_t) payload_length);
     metadata_length = (tsk_size_t) total_length;
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, 0);
     CU_ASSERT_PTR_EQUAL(json, (const char *) bytes + header_length);
@@ -165,7 +165,7 @@ test_json_binary_metadata_get_blob(void)
     set_u64_le(bytes + 13, (uint64_t) payload_length);
     memcpy(bytes + header_length + json_length, empty_payload, payload_length);
     metadata_length = (tsk_size_t) total_length;
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, 0);
     CU_ASSERT_PTR_EQUAL(json, (const char *) bytes + header_length);
@@ -179,43 +179,63 @@ test_json_binary_metadata_get_blob(void)
     json = NULL;
     json_buffer_length = 0;
     metadata_length = header_length - 1;
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_FORMAT);
 
     metadata_length = (tsk_size_t) total_length;
     bytes[0] = 'X';
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_FORMAT);
     bytes[0] = 'J';
 
     bytes[4] = 2;
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_VERSION_TOO_NEW);
     bytes[4] = 1;
 
     metadata_length = (tsk_size_t)(total_length - 1);
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_FORMAT);
 
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         NULL, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_BAD_PARAM_VALUE);
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, NULL, &json_buffer_length, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_BAD_PARAM_VALUE);
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, NULL, &blob, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_BAD_PARAM_VALUE);
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, NULL, &blob_length);
     CU_ASSERT_EQUAL(ret, TSK_ERR_BAD_PARAM_VALUE);
-    ret = tsk_json_binary_metadata_get_blob(
+    ret = tsk_json_struct_metadata_get_blob(
         metadata, metadata_length, &json, &json_buffer_length, &blob, NULL);
     CU_ASSERT_EQUAL(ret, TSK_ERR_BAD_PARAM_VALUE);
+
+    memset(metadata, 0, sizeof(metadata));
+    bytes[0] = 'J';
+    bytes[1] = 'B';
+    bytes[2] = 'L';
+    bytes[3] = 'B';
+    bytes[4] = 1;
+    metadata_length = (tsk_size_t) header_length;
+
+    set_u64_le(bytes + 5, UINT64_MAX - (uint64_t) header_length + 1);
+    set_u64_le(bytes + 13, 0);
+    ret = tsk_json_struct_metadata_get_blob(
+        metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
+    CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_FORMAT);
+
+    set_u64_le(bytes + 5, 8);
+    set_u64_le(bytes + 13, UINT64_MAX - (uint64_t)(header_length + 8) + 1);
+    ret = tsk_json_struct_metadata_get_blob(
+        metadata, metadata_length, &json, &json_buffer_length, &blob, &blob_length);
+    CU_ASSERT_EQUAL(ret, TSK_ERR_FILE_FORMAT);
 }
 
 static void
@@ -788,7 +808,7 @@ main(int argc, char **argv)
         { "test_strerror", test_strerror },
         { "test_strerror_kastore", test_strerror_kastore },
         { "test_generate_uuid", test_generate_uuid },
-        { "test_json_binary_metadata_get_blob", test_json_binary_metadata_get_blob },
+        { "test_json_struct_metadata_get_blob", test_json_struct_metadata_get_blob },
         { "test_double_round", test_double_round },
         { "test_blkalloc", test_blkalloc },
         { "test_unknown_time", test_unknown_time },